requirements.txt
Django==5.1.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn== 22.0.0
drf-yasg==1.21.5
python-decouple==3.7
setuptools== 75.1.0
